refactor project hooks
